refactor: Update timestamp type to int64_t in config.h and transaction.h

+++ src/common/config.h
@@ -50,7 +50,7 @@ using lsn_t = int32_t;         // log sequence number type
 using slot_offset_t = size_t;  // slot offset type
 using oid_t = uint16_t;
 using timestamp_t =
-    int32_t;  // timestamp type, used for transaction concurrency
+    int64_t;  // timestamp type, used for transaction concurrency
 
 // log file
 static const std::string LOG_FILE_NAME = "db.log";
